replace getServerCount by getting int value from Long, instead of getting the exact number

+++ src/main/java/dev/sergivos/toastr/backend/redis/RedisManager.java
@@ -324,7 +324,7 @@ public class RedisManager {
      */
     public int getServerCount(String server) {
         try(final Jedis jedis = getConnection()) {
-            return Math.toIntExact(jedis.scard("server:" + server));
+            return jedis.scard("server:" + server).intValue();
         }
     }
